dividing rest controller into interface and implementation
