+++ src/main/java/com/naruhin/springbootexamplehillelhw5/web/AddressRestController.java
@@ -1,56 +1,30 @@
 package com.naruhin.springbootexamplehillelhw5.web;
 
-import com.naruhin.springbootexamplehillelhw5.config.AddressMapper;
-import com.naruhin.springbootexamplehillelhw5.domain.Address;
 import com.naruhin.springbootexamplehillelhw5.dto.AddressDTO;
-import com.naruhin.springbootexamplehillelhw5.service.AddressService;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.web.bind.annotation.*;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import io.swagger.v3.oas.annotations.tags.Tag;
 
-import java.util.Collection;
 import java.util.List;
 
-@RestController
-@RequestMapping(value = "/api", produces = MediaType.APPLICATION_JSON_VALUE)
-public class AddressRestController {
+@Tag(name = "Address", description = "Car API")
 
-    private final AddressService service;
+public interface AddressRestController {
 
-    public AddressRestController(AddressService service) {
-        this.service = service;
-    }
+    @Operation(summary = "Add a new address", description = "endpoint for creating an entity", tags = {"Address"})
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "201", description = "Address created"),
+            @ApiResponse(responseCode = "400", description = "Invalid input"),
+            @ApiResponse(responseCode = "409", description = "Address already exists")})
+    AddressDTO saveAddress(AddressDTO addressDTO);
 
-    //Операция сохранения адресов в базу данных
-    @PostMapping("/addresses")
-    @ResponseStatus(HttpStatus.CREATED)
-    public AddressDTO saveAddress(@RequestBody AddressDTO addressDTO) {
-        Address address = AddressMapper.INSTANCE.toAddress(addressDTO);
-        return AddressMapper.INSTANCE.toAddressDto(service.saveAddress(address));
-    }
+    @Operation(summary = "Get list of addresses", description = "endpoint to get list of entities", tags = {"Address"})
+    List<AddressDTO> getAllAddresses();
 
-    //Получение списка адресов
-    @GetMapping("/addresses")
-    @ResponseStatus(HttpStatus.OK)
-    public List<AddressDTO> getAllAddresses() {
-        Collection<Address> addresses = service.getAllAddresses();
-        return AddressMapper.INSTANCE.map((List<Address>) addresses);
-    }
+    @Operation(summary = "Remove all addresses", description = "remove all entities", tags = {"Address"})
+    void removeAllAddresses();
 
-    //Удаление всех адресов
-    @DeleteMapping("/addresses")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void removeAllAddresses() {
-        service.removeAllAddresses();
-    }
-
-    //Обновление адреса
-    @PutMapping("/addresses/{id}")
-    @ResponseStatus(HttpStatus.OK)
-    public AddressDTO updateAddress(@PathVariable("id") long id, @RequestBody AddressDTO addressDTO) {
-        Address address = AddressMapper.INSTANCE.toAddress(addressDTO);
-        return AddressMapper.INSTANCE.toAddressDto(service.updateAddress(id,address));
-
-    }
-
+    @Operation(summary = "Update a cars", description = "endpoint to update entity", tags = {"Address"})
+    AddressDTO updateAddress(long id, AddressDTO addressDTO);
 }
