+++ src/main/java/com/naruhin/springbootexamplehillelhw5/web/CarRestController.java
@@ -1,94 +1,42 @@
 package com.naruhin.springbootexamplehillelhw5.web;
 
-import com.naruhin.springbootexamplehillelhw5.config.CarMapper;
-import com.naruhin.springbootexamplehillelhw5.config.ManufacturerMapper;
-import com.naruhin.springbootexamplehillelhw5.domain.Car;
-import com.naruhin.springbootexamplehillelhw5.domain.Manufacturer;
 import com.naruhin.springbootexamplehillelhw5.dto.CarDTO;
-import com.naruhin.springbootexamplehillelhw5.service.CarService;
-import com.naruhin.springbootexamplehillelhw5.service.ManufacturerService;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.web.bind.annotation.*;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import io.swagger.v3.oas.annotations.tags.Tag;
 
 import java.util.Collection;
-import java.util.List;
 
-@RestController
-@RequestMapping(value = "/api", produces = MediaType.APPLICATION_JSON_VALUE)
-public class CarRestController {
+@Tag(name = "Car", description = "Car API")
 
-    private final CarService carService;
-    private final ManufacturerService manufacturerService;
+public interface CarRestController {
 
-    public CarRestController(CarService carService, ManufacturerService manufacturerService) {
-        this.carService = carService;
-        this.manufacturerService = manufacturerService;
-    }
+    @Operation(summary = "Add a new car", description = "endpoint for creating an entity", tags = {"Car"})
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "201", description = "Car created"),
+            @ApiResponse(responseCode = "400", description = "Invalid input"),
+            @ApiResponse(responseCode = "409", description = "Car already exists")})
+    CarDTO saveCar(CarDTO carDTO, long manufacturerId);
 
-    //Операция сохранения машины в базу данных
-    @PostMapping("/cars/manufacturers/{manufacturerId}")
-    @ResponseStatus(HttpStatus.CREATED)
-    public CarDTO saveCar(@RequestBody CarDTO carDTO, @PathVariable long manufacturerId) {
-        Manufacturer manufacturer = manufacturerService.getManufacturerByID(manufacturerId);
-        Car car = CarMapper.INSTANCE.toCar(carDTO);
-        carDTO.setManufacturer(ManufacturerMapper.INSTANCE.toManufacturerDto(manufacturer));
-        carService.saveCar(car,manufacturerId);
-        return CarMapper.INSTANCE.toCarDto(car);
-    }
+    @Operation(summary = "Get list of cars", description = "endpoint to get list of entities", tags = {"Car"})
+    Collection<CarDTO> getAllCars();
 
-    //Получение списка машин
-    @GetMapping("/cars")
-    @ResponseStatus(HttpStatus.OK)
-    public Collection<CarDTO> getAllCars() {
-        List<Car> cars = (List<Car>) carService.getAllCars();
-        return CarMapper.INSTANCE.map(cars);
-    }
+    @Operation(summary = "Get car by ID", description = "endpoint to get an entity by ID", tags = {"Car"})
+    CarDTO getCarById(long id);
 
-    //Получения машины по id
-    @GetMapping("/cars/{id}")
-    @ResponseStatus(HttpStatus.OK)
-    public CarDTO getCarById(@PathVariable long id) {
-        Car car = carService.getCarById(id);
-        return CarMapper.INSTANCE.toCarDto(car);
-    }
+    @Operation(summary = "Update car", description = "endpoint for updating an entity", tags = {"Car"})
+    CarDTO updateCar(long id, CarDTO carDTO);
 
-    //Обновление машины
-    @PutMapping("/cars/{id}")
-    @ResponseStatus(HttpStatus.OK)
-    public CarDTO updateCar(@PathVariable("id") long id, @RequestBody CarDTO carDTO) {
-        Car car = CarMapper.INSTANCE.toCar(carDTO);
-        return  CarMapper.INSTANCE.toCarDto(carService.updateCar(id, car));
-    }
+    @Operation(summary = "Update car's dealer", description = "endpoint for updating dealer", tags = {"Car"})
+    CarDTO updateCarDealer(long carId, long dealerId);
 
+    @Operation(summary = "Update car's service station", description = "endpoint for updating service station", tags = {"Car"})
+    CarDTO updateCarServiceStation(long carId, long serviceStationId);
 
-    @PatchMapping("/cars/{carId}/dealers/{dealerId}")
-    @ResponseStatus(HttpStatus.OK)
-    public CarDTO updateCarDealer(@PathVariable long carId, @PathVariable long dealerId){
-        carService.updateCarDealer(carId,dealerId);
-        return CarMapper.INSTANCE.toCarDto(carService.getCarById(carId));
-    }
+    @Operation(summary = "Delete all cars", description = "endpoint for deleting all entities", tags = {"Car"})
+    void removeAllCars();
 
-    @PatchMapping("/cars/{carId}/service_stations/{serviceStationId}")
-    @ResponseStatus(HttpStatus.OK)
-    public CarDTO updateCarServiceStation(@PathVariable long carId, @PathVariable long serviceStationId){
-        carService.updateCarServiceStation(carId,serviceStationId);
-        return CarMapper.INSTANCE.toCarDto(carService.getCarById(carId));
-    }
-
-
-    //Удаление всех машин
-    @DeleteMapping("/cars")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void removeAllCars() {
-        carService.removeAllCars();
-    }
-
-    //Удаление по id
-    @DeleteMapping("/cars/{id}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void removeCarsById(@PathVariable long id) {
-        carService.removeCarsById(id);
-    }
-
+    @Operation(summary = "Delete car by ID", description = "endpoint for deleting an entity by ID", tags = {"Car"})
+    void removeCarsById(long id);
 }
