+++ src/main/java/com/naruhin/springbootexamplehillelhw5/web/ManufacturerRestController.java
@@ -1,62 +1,30 @@
 package com.naruhin.springbootexamplehillelhw5.web;
 
-import com.naruhin.springbootexamplehillelhw5.config.AddressMapper;
-import com.naruhin.springbootexamplehillelhw5.config.ManufacturerMapper;
-import com.naruhin.springbootexamplehillelhw5.domain.Address;
-import com.naruhin.springbootexamplehillelhw5.domain.Manufacturer;
 import com.naruhin.springbootexamplehillelhw5.dto.ManufacturerDTO;
-import com.naruhin.springbootexamplehillelhw5.service.AddressService;
-import com.naruhin.springbootexamplehillelhw5.service.ManufacturerService;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.web.bind.annotation.*;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import io.swagger.v3.oas.annotations.tags.Tag;
 
 import java.util.Collection;
-import java.util.List;
 
-@RestController
-@RequestMapping(value = "/api", produces = MediaType.APPLICATION_JSON_VALUE)
-public class ManufacturerRestController {
-    private final AddressService addressService;
-    private final ManufacturerService manufacturerService;
+@Tag(name = "Manufacturer", description = "Car API")
 
-    public ManufacturerRestController(AddressService addressService, ManufacturerService manufacturerService) {
-        this.addressService = addressService;
-        this.manufacturerService = manufacturerService;
-    }
+public interface ManufacturerRestController {
+    @Operation(summary = "Add a new manufacturer", description = "endpoint for creating an entity", tags = {"Manufacturer"})
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "201", description = "Manufacturer created"),
+            @ApiResponse(responseCode = "400", description = "Invalid input"),
+            @ApiResponse(responseCode = "409", description = "Manufacturer already exists")})
+    ManufacturerDTO saveManufacturer(ManufacturerDTO manufacturerDTO, long addressId);
 
-    //Операция сохранения производителя в базу данных
-    @PostMapping("/manufacturers/{addressId}")
-    @ResponseStatus(HttpStatus.CREATED)
-    public ManufacturerDTO saveManufacturer(@RequestBody ManufacturerDTO manufacturerDTO, @PathVariable long addressId) {
-        Address address = addressService.getAddressByID(addressId);
-        Manufacturer manufacturer = ManufacturerMapper.INSTANCE.toManufacturer(manufacturerDTO);
+    @Operation(summary = "Get list of manufacturers", description = "endpoint for getting list of entities", tags = {"Manufacturer"})
+    Collection<ManufacturerDTO> getAllManufacturers();
 
-        manufacturerDTO.setAddress(AddressMapper.INSTANCE.toAddressDto(address));
+    @Operation(summary = "Remove all manufacturers", description = "endpoint for deleting an entity", tags = {"Manufacturer"})
+    void removeAllManufacturers();
 
-        return ManufacturerMapper.INSTANCE.toManufacturerDto(manufacturerService.saveManufacturer(manufacturer,addressId));
-    }
+    @Operation(summary = "Update a manufacturer", description = "endpoint for updating an entity", tags = {"Manufacturer"})
+    ManufacturerDTO updateManufacturer(long id, ManufacturerDTO manufacturerDTO);
 
-    //Получение списка производителей
-    @GetMapping("/manufacturers")
-    @ResponseStatus(HttpStatus.OK)
-    public Collection<ManufacturerDTO> getAllManufacturers() {
-        Collection<Manufacturer> manufacturers = manufacturerService.getAllManufacturers();
-        return ManufacturerMapper.INSTANCE.map((List<Manufacturer>) manufacturers);
-    }
-
-    //Удаление всех производителей
-    @DeleteMapping("/manufacturers")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void removeAllManufacturers() {
-        manufacturerService.removeAllManufacturers();
-    }
-
-    //Обновление производителя
-    @PutMapping("/manufacturers/{id}")
-    @ResponseStatus(HttpStatus.OK)
-    public ManufacturerDTO updateManufacturer(@PathVariable("id") long id, @RequestBody ManufacturerDTO manufacturerDTO) {
-        Manufacturer manufacturer = ManufacturerMapper.INSTANCE.toManufacturer(manufacturerDTO);
-        return  ManufacturerMapper.INSTANCE.toManufacturerDto(manufacturerService.updateManufacturer(id, manufacturer));
-    }
 }
